feat: add regional NWP models for international cities, fix °C/°F conversion and Dallas station
- Add local NWP models (ICON, AROME, KMA, GEM) as 3rd ensemble source for international cities
- Fix critical °C→°F bucket conversion for international Polymarket markets
- Fix Dallas station: KDFW (DFW) → KDAL (Love Field) to match Polymarket resolution
- Update probability.py type hints (int→float for converted bucket bounds)
- Add 20 location validation tests (station/coordinate/unit alignment with Polymarket)
- Update README for 14 cities (6 US + 8 international), regional NWP models, corrected station mappings

+++ weather/probability.py
@@ -522,8 +522,8 @@ def _weather_sigma_multiplier(weather_data: dict, metric: str = "high") -> float
 
 def estimate_bucket_probability(
     forecast_temp: float,
-    bucket_low: int,
-    bucket_high: int,
+    bucket_low: float,
+    bucket_high: float,
     forecast_date: str,
     apply_seasonal: bool = True,
     location: str = "",
@@ -672,8 +672,8 @@ def _tz_from_lon(lon: float) -> str:
 
 def estimate_bucket_probability_with_obs(
     forecast_temp: float,
-    bucket_low: int,
-    bucket_high: int,
+    bucket_low: float,
+    bucket_high: float,
     forecast_date: str,
     obs_data: dict | None = None,
     metric: str = "high",
